refactor(accounts): Use a single file for all requests schemas
